@@ -3,17 +3,21 @@
   "private": true,
   "version": "0.0.0",
   "main": "./out/main/main.js",
+  "type": "module",
   "scripts": {
     "dev": "electron-vite dev -w",
     "build": "tsc && electron-vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "electron-vite preview"
+    "lint": "eslint .",
+    "preview": "electron-vite preview",
+    "pack": "electron-builder --dir",
+    "dist": "electron-builder"
   },
   "dependencies": {
     "@hookform/resolvers": "^3.3.4",
     "@types/eslint": "^8.56.9",
     "classnames": "^2.5.1",
     "prettier": "^3.2.5",
+    "prettier-eslint": "^16.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.3",
@@ -21,6 +25,8 @@
     "yup": "^1.4.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.1.1",
+    "@types/node": "^20.12.7",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^7.7.0",
@@ -28,17 +34,34 @@
     "@vitejs/plugin-react": "^4.2.1",
     "add": "^2.0.6",
     "electron": "^30.0.0",
+    "electron-builder": "^24.13.3",
     "electron-vite": "^2.1.0",
-    "eslint": "^9.0.0",
+    "eslint": "^9.1.0",
     "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
     "sass": "^1.75.0",
-    "typescript": "^5.2.2",
+    "typescript": "^5.4.5",
+    "typescript-eslint": "^7.7.0",
     "vite": "^5.2.0",
     "vite-plugin-eslint": "^1.8.1",
+    "vite-plugin-eslint2": "^4.4.0",
     "yarn": "^1.22.22"
+  },
+  "build": {
+    "appId": "com.cool_encryption.app",
+    "win": {
+      "target": [
+        {
+          "target": "nsis",
+          "arch": [
+            "x64",
+            "ia32"
+          ]
+        }
+      ]
+    }
   }
 }
